% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prior.R
\name{build_prior}
\alias{build_prior}
\alias{build_GIG}
\alias{build_GBP}
\title{Construct a prior specification for GBASS}
\usage{
build_prior(
  type = c("GIG", "GBP"),
  p,
  a,
  b,
  lower_bound = NULL,
  prop_sigma = NULL,
  adapt = NULL,
  adapt_delay = NULL,
  adapt_thin = NULL
)

build_GIG(
  p,
  a,
  b,
  lower_bound = NULL,
  prop_sigma = NULL,
  adapt = NULL,
  adapt_delay = NULL,
  adapt_thin = NULL
)

build_GBP(
  p,
  a,
  b,
  lower_bound = NULL,
  prop_sigma = NULL,
  adapt = NULL,
  adapt_delay = NULL,
  adapt_thin = NULL
)
}
\arguments{
\item{type}{"GIG" or "GBP"}

\item{p, a, b}{prior hyperparameters}

\item{lower_bound}{optional lower bound}

\item{prop_sigma}{optional proposal sd (log scale)}

\item{adapt}{logical; use adaptive MH?}

\item{adapt_delay}{delay before adapting}

\item{adapt_thin}{thinning for adaptation updates}
}
\description{
Construct a prior specification for GBASS
}
