% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{Concrete Data}
\format{
Data frame formatting information
}
\source{
add link here
}
\usage{
concrete
}
\description{
A data set consisting of measurements of concrete, fill this in later.
}
\keyword{datasets}
