% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpon.R
\name{dmpon}
\alias{dmpon}
\title{Un-normalized density of the "monomial perturbation of normal (mpon)" distribution}
\usage{
dmpon(x, alpha, gamma, mu, log = FALSE)
}
\arguments{
\item{x}{locations for density evaluation}

\item{alpha}{parameter}

\item{gamma}{parameter}

\item{mu}{parameter}
}
\value{
density of the mpon distribution
}
\description{
A function to evaluate the unnormalized density of the mpon distribution.
}
\details{
Evaluates the (unnormalized) density function f(x) = x^(alpha-1)exp(-gamma*(x-mu)^2)
}
\examples{
n <- 10000
alpha <- 5
gamma <- 1
mu <- 2
y <- rep(NA, n)
for(i in 1:n){
  y[i] <- rmpon(1, alpha, gamma, mu)
}
hist(y, breaks=30, freq=F)
c <- integrate(dmpon, lower=0, upper=20, alpha=alpha, gamma=gamma, mu=mu)$value
curve(dmpon(x, alpha, gamma, mu)/c, add=TRUE, lwd=2, col='blue')
}
