% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{funcs}
\alias{funcs}
\alias{ff1}
\alias{ff2}
\alias{ff3}
\alias{ff4}
\alias{ff5}
\title{Bivariate functions for Denison, Mallick and Smith}
\usage{
ff1(x)

ff2(x)

ff3(x)

ff4(x)

ff5(x)
}
\arguments{
\item{x}{a vector of inputs. length(x) must be at least 2, but only the first two entries are used.}
}
\description{
Bivariate functions for Denison, Mallick and Smith
}
\details{
The five test functions from Denison, Mallick and Smith (1998) are referred to as
\enumerate{
\item Simple
\item Radial
\item Harmonic
\item Additive
\item Complex interaction
}
}
\examples{
n <- 100 #Number of observations
p <- 4   #Number of variables (beyond p = 2, variables are inert)
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, ff1)
}
