% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBASS-package.R
\docType{package}
\name{GBASS-package}
\alias{GBASS}
\alias{GBASS-package}
\title{GBASS: Generalized Bayesian Adaptive Smoothing Splines}
\description{
Bayesian nonlinear regression under a range of likelihood models using generalized Bayesian adaptive smoothing splines. Robust regression with Student's t likelihoods, quantile regression, and related latent-scale models are included as special cases.
}
\author{
\strong{Maintainer}: Kellin Rumsey \email{knrumsey@lanl.gov}

}
