% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gm2bm}
\alias{gm2bm}
\title{Convert GBASS to BASS}
\usage{
gm2bm(gm)
}
\arguments{
\item{gm}{an object of class gbass}
}
\description{
This function converts an object of class \code{gbass} to an object of class \code{bass}. This is convenient for taking advantage of
the sobol decomposition functions available in the BASS package.
}
\examples{
#The following are equivalent
n <- 100 #Number of observations
p <- 4   #Number of variables (beyond p = 2, variables are inert)
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, ff1)
gm <- gbass(X, Y, nmcmc=1000, nburn=901)
bm <- gm2bm(gm)
sob <- sobol(bm)
plot(sob)

}
