% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_old.R
\name{hbassold}
\alias{hbassold}
\title{bass_shoe or hbass Wrapper}
\usage{
hbassold(
  X,
  y,
  prop_sigma,
  w_prior = list(type = "GBP", p = 1, a = 1/2, b = 1/2, prop_sigma = 0.5),
  control = list(),
  ...
)
}
\arguments{
\item{prop_sigma}{proposal sd for v_i}
}
\description{
gbmars() wrapper for horseshoe likelihood
}
\examples{
foo <- 1 + 1

}
