% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbass.R
\name{plot.gbass}
\alias{plot.gbass}
\title{Plot method for gbass objects}
\usage{
\method{plot}{gbass}(x, ...)
}
\arguments{
\item{x}{A fitted object of class \code{"gbass"}.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot method for gbass objects
}
