% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbass.R
\name{print.gbass}
\alias{print.gbass}
\title{Print a gbass object}
\usage{
\method{print}{gbass}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"gbass"}.}

\item{...}{Unused.}
}
\description{
Print a gbass object
}
