% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_old.R
\name{qbassold}
\alias{qbassold}
\title{quantile regression bass wrapper}
\usage{
qbassold(
  X,
  y,
  q = 0.5,
  w_prior = list(type = "GIG", p = 0.001, a = 0.001, b = 0, prop_sigma = 0.2),
  control = list(),
  ...
)
}
\arguments{
\item{q}{quantile of interest (default 0.5)}
}
\description{
gbmars() wrapper for quantile regression
}
\examples{
foo <- 1 + 1

}
