% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMHN.R
\name{rMHN_gamma_positive_N}
\alias{rMHN_gamma_positive_N}
\title{Gamma_Positive_N}
\usage{
rMHN_gamma_positive_N(n, alpha, beta, gamma)
}
\description{
Gamma_Positive_N
}
\examples{
ss=Gamma_Positive_N(N=1000, alpha=20, beta=1, gamma=1)
hist(ss$sample,prob=TRUE, col="grey")
lines(density(ss$sample))
}
