% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpon.R
\name{rmpon}
\alias{rmpon}
\title{Random generator for the "monomial perturbation of normal" distribution}
\usage{
rmpon(n = 1, alpha, gamma, mu, max_steps = 10000, tol = 1e-04)
}
\arguments{
\item{n}{currently, n = 1. Changing this argument has no effect on the code.}

\item{alpha}{parameter}

\item{gamma}{parameter}

\item{mu}{parameter}

\item{max_steps}{maximum number of steps for Newton Raphson}

\item{tol}{tolerance for Newton Raphson}
}
\value{
a random draw from the mpon distribution
}
\description{
A function to generate from the mpon distribution. Uses a uniformly bounded rejection sampling scheme, adapted from Devroye (2014).
}
\details{
Generates a RV with (unnormalized) density function f(x) = x^(alpha-1)exp(-gamma*(x-mu)^2)
}
\examples{
n <- 10000
alpha <- 5
gamma <- 1
mu <- 2
y <- rep(NA, n)
for(i in 1:n){
  y[i] <- rmpon(1, alpha, gamma, mu)
}
hist(y, breaks=30, freq=F)
c <- integrate(dmpon, lower=0, upper=20, alpha=alpha, gamma=gamma, mu=mu)$value
curve(dmpon(x, alpha, gamma, mu)/c, add=TRUE, lwd=2, col='blue')
}
