% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpon.R
\name{rmpon_sun}
\alias{rmpon_sun}
\title{Random generator for the "monomial perturbation of normal" distribution}
\usage{
rmpon_sun(n = 1, alpha, gamma, mu, max_steps = 1e+06)
}
\arguments{
\item{n}{currently, n = 1. Changing this argument has no effect on the code.}

\item{alpha}{parameter}

\item{gamma}{parameter}

\item{mu}{parameter}

\item{max_steps}{maximum number of steps for rejection sampler. Returns error if met.}
}
\value{
a random draw from the mpon distribution
}
\description{
A function to generate from the mpon distribution. This is an alternative approach based on Sun et al (2021). Assumes alpha > 1.
(depreciated? See modified half normal distribution)
}
\details{
Generates a RV with (unnormalized) density function f(x) = x^(alpha-1)exp(-gamma*(x-mu)^2)
}
\examples{
n <- 10000
alpha <- 5
gamma <- 1
mu <- 2
y <- rep(NA, n)
for(i in 1:n){
  y[i] <- rmpon(1, alpha, gamma, mu)
}
hist(y, breaks=30, freq=F)
c <- integrate(dmpon, lower=0, upper=20, alpha=alpha, gamma=gamma, mu=mu)$value
curve(dmpon(x, alpha, gamma, mu)/c, add=TRUE, lwd=2, col='blue')
}
