% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_old.R
\name{tbassold}
\alias{tbassold}
\title{T BASS Wrapper}
\usage{
tbassold(
  X,
  y,
  df = 1,
  w_prior = list(type = "GIG", p = 0.001, a = 0.001, b = 0),
  ...
)
}
\arguments{
\item{df}{degrees of freedom, default=1 (Cauchy)}
}
\description{
gbmars() wrapper for t likelihoods
}
\examples{
foo <- 1 + 1

}
