% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{build_prior}
\alias{build_prior}
\title{Build Prior}
\usage{
build_prior(type, pars, prop_sigma = NULL)
}
\arguments{
\item{type}{either "GIG" or "GBP"}

\item{pars}{a vector (p, a, b) of parameters}

\item{prop_sigma}{the proposal standard deviation (only used for GBP or when symmetry parameter (beta) is non-zero)}
}
\description{
This function provides a short cut for specifying GIG or GBP priors. See main help page of gbass() for more details.
}
\details{
Shortcut for specifying priors for the gbass() function
}
\examples{
#The following are equivalent
w_prior <- list(type="GBP", p=1, a=1/2, b=1/2, prop_sigma = 0.2)
w_prior <- build_prior("GBP", c(1, 1/2, 1/2), 0.2)

}
