% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prior.R
\name{build_prior}
\alias{build_prior}
\alias{build_GIG}
\alias{build_GBP}
\title{Build GIG and GBP priors}
\usage{
build_prior(type, p, a, b, prop_sigma = NULL)

build_GIG(p, a, b, prop_sigma = NULL)

build_GBP(p, a, b, prop_sigma = NULL)
}
\arguments{
\item{type}{either "GIG" or "GBP"}

\item{prop_sigma}{the proposal standard deviation (only used for GBP or when symmetry parameter (beta) is non-zero)}

\item{pars}{a vector (p, a, b) of parameters}
}
\description{
Build Prior
}
\details{
This function provides a short cut for specifying GIG or GBP priors. See main help page of gbass() for more details.

Shortcut for specifying priors for the gbass() function. prop_sigma may not be necessary for GIG prior.
}
\examples{
#The following are equivalent
w_prior <- list(type="GBP", p=1, a=1/2, b=1/2, prop_sigma = 0.2)
w_prior <- build_prior("GBP", c(1, 1/2, 1/2), 0.2)
# The following are equivalent
w_prior <- list(type="GBP", p=1, a=1/2, b=1/2, prop_sigma = 0.2)
w_prior <- build_prior("GBP", 1, 1/2, 1/2, 0.2)
w_prior <- build_GBP(1, 1/2, 1/2, 0.2)
}
