% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhbass.R
\name{fhbass}
\alias{fhbass}
\title{Fay-Herriot BMARS}
\usage{
fhbass(
  X,
  y,
  d,
  maxInt = 3,
  maxBasis = 1000,
  npart = NULL,
  nmcmc = 10000,
  nburn = 9001,
  thin = 1,
  moveProbs = rep(1/3, 3),
  gamma_v_scale = 100,
  gamma_w_scale = NULL,
  fh_sample_var = NULL,
  a_tau = 1/2,
  b_tau = NULL,
  a_lambda = 1,
  b_lambda = 1,
  m_beta = 0,
  s_beta = 0,
  scale = 1,
  Iw0 = rep(1, maxInt),
  Zw0 = rep(1, ncol(X)),
  verbose = TRUE
)
}
\arguments{
\item{X}{an Nxp matrix of predictor variables.}

\item{y}{an Nx1 matrix (or vector) of response values.}

\item{d}{A vector of delta values (measured variance at each location). See details.}

\item{maxInt}{integer for maximum degree of interaction in spline basis functions. Defaults to the number of predictors, which could result in overfitting.}

\item{maxBasis}{maximum number of basis functions. This should probably only be altered if you run out of memory.}

\item{npart}{of non-zero points in a basis function. If the response is functional, this refers only to the portion of the basis function coming from the non-functional predictors. Defaults to 20 or 0.1 times the number of observations, whichever is smaller.}

\item{nmcmc}{number of mcmc iterations}

\item{nburn}{number of burn-in samples}

\item{thin}{thinning for mcmc}

\item{moveProbs}{a vector defining the probabilities for (i) birth (ii) death and (iii) mutation. Default is rep(1/3,3).}

\item{gamma_v_scale}{scale of the half-Cauchy prior for \code{gamma_v}.}

\item{gamma_w_scale}{scale of the half-Cauchy prior for \code{gamma_w}.}

\item{fh_sample_var}{a prior guess for the classical Fay-Herriot "sampling error variance". Used to generate prior for \code{gamma_w}.}

\item{a_tau}{prior for tau}

\item{b_tau}{prior for tyau}

\item{a_lambda}{prior for lambda}

\item{b_lambda}{prior for lambda}

\item{m_beta}{prior for beta}

\item{s_beta}{prior for beta}

\item{scale}{fixed variance parameter. default is one.}

\item{Iw0}{vector of nominal weights for degree of interaction, used in generating candidate basis functions. Should have length equal to Jmax and have positive entries.}

\item{Zw0}{vector of nominal weights for variable selection, used in generating candidate basis functions. Should have length equal to ncol(X) and have positive entries.}

\item{verbose}{Logical. Should gbass print completion status? Default TRUE}
}
\description{
A version of BMARS which attempts to replicate the Fay-Herriot model.
}
\details{
The prior \code{v_prior_fh} should be a named list with named fields \code{df} (a scalar) and \code{m, vhat} (n-vectors).
}
\examples{
fm <- function (x){
  x[1]^2 + x[1] * x[2] +  x[2]^3 / 9
}

fv <- function (x){
  1.3356 * (1.5 * (1 - x[1]) + exp(2 * x[1] - 1) * sin(3 * pi *
    (x[1] - 0.6)^2) + exp(3 * (x[2] - 0.5)) * sin(4 * pi * (x[2] -
     0.9)^2))
}

# GENERATE DATA
n <- 50
sigma_v <- 0.1
X <- lhs::randomLHS(n, 2)
mu <- apply(X, 1, fm)
mu <- mu / sd(mu)
N_pop <- 1 + round(apply(X, 1, fv))
d <- 1/N_pop
y <- rnorm(n, mu, sqrt(d)) + rnorm(n, 0, sigma_v)

# Fit model
fit <- fhbass(X, y, d)

}
