% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhbass.R
\name{fhbass}
\alias{fhbass}
\title{Fay-Harris BMARS}
\usage{
fhbass(
  X,
  y,
  w_prior = list(type = "GIG", p = 0, a = 0, b = 0),
  v_prior_fh = list(type = "GIG", p = -15, a = 0, b = 30),
  maxInt = 3,
  maxBasis = 1000,
  npart = NULL,
  nmcmc = 10000,
  nburn = 9001,
  thin = 1,
  moveProbs = rep(1/3, 3),
  a_tau = 1/2,
  b_tau = NULL,
  a_lambda = 1,
  b_lambda = 1,
  m_beta = 0,
  s_beta = 0,
  scale = 1,
  Iw0 = rep(1, maxInt),
  Zw0 = rep(1, ncol(X)),
  verbose = TRUE
)
}
\arguments{
\item{X}{an Nxp matrix of predictor variables.}

\item{y}{an Nx1 matrix (or vector) of response values.}

\item{w_prior}{a named list specifying the prior for the global variance component. See details.}

\item{v_prior_fh}{a named list specifying the (not-shared) prior for the local variance components. See details.}

\item{maxInt}{integer for maximum degree of interaction in spline basis functions. Defaults to the number of predictors, which could result in overfitting.}

\item{maxBasis}{maximum number of basis functions. This should probably only be altered if you run out of memory.}

\item{npart}{of non-zero points in a basis function. If the response is functional, this refers only to the portion of the basis function coming from the non-functional predictors. Defaults to 20 or 0.1 times the number of observations, whichever is smaller.}

\item{nmcmc}{number of mcmc iterations}

\item{nburn}{number of burn-in samples}

\item{thin}{thinning for mcmc}

\item{moveProbs}{a vector defining the probabilities for (i) birth (ii) death and (iii) mutation. Default is rep(1/3,3).}

\item{a_tau}{prior for tau}

\item{b_tau}{prior for tyau}

\item{a_lambda}{prior for lambda}

\item{b_lambda}{prior for lambda}

\item{m_beta}{prior for beta}

\item{s_beta}{prior for beta}

\item{scale}{fixed variance parameter. default is one.}

\item{Iw0}{vector of nominal weights for degree of interaction, used in generating candidate basis functions. Should have length equal to Jmax and have positive entries.}

\item{Zw0}{vector of nominal weights for variable selection, used in generating candidate basis functions. Should have length equal to ncol(X) and have positive entries.}

\item{verbose}{Logical. Should gbass print completion status? Default TRUE}
}
\description{
A version of BMARS which attempts to replicate the Fay-Harris model.
}
\details{
The prior \code{v_prior_fh} should be a named list with named fields \code{df} (a scalar) and \code{m, vhat} (n-vectors).
}
\examples{
n <- 100 #Number of observations
p <- 4   #Number of variables (beyond p = 2, variables are inert)
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, ff1)
mod <- gbass(X, y, nmcmc=1000, nburn=901, thin=2)

}
