% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gbass2bass}
\alias{gbass2bass}
\alias{gm2bm}
\title{Convert GBASS object to BASS-like object}
\usage{
gbass2bass(gm)

gm2bm(gm)
}
\arguments{
\item{gm}{An object of class \code{gbass}.}
}
\value{
An object with class \code{c("bass", "gbass")}.
}
\description{
Converts an object of class \code{gbass} to an object with class \code{bass},
so that downstream BASS utilities such as Sobol decomposition can be used.
}
