% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hbass}
\alias{hbass}
\alias{bass_shoe}
\title{HBASS - Bayesian MARS with a Horseshoe (or Strawderman-Berger) likelihood}
\usage{
hbass(X, y, likelihood = "h", ...)
}
\arguments{
\item{X}{an Nxp matrix of predictor variables.}

\item{y}{an Nx1 matrix (or vector) of response values.}

\item{likelihood}{default "h". Also accepts "sb" for Strawderman-Berger}
}
\description{
HBASS - Bayesian MARS with a Horseshoe (or Strawderman-Berger) likelihood
}
\details{
The Horseshoe and Strawderman-Berger likelihoods can be unified under the class of
Generalized Beta Prime distributions. This function provides a wrapper for these likelilhoods. For additional flexibility, use gbass().
These distributions have most of their mass near zero but with heavy tails, making this an interesting choice for datasets with potential outliers.
}
