% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwbass.R
\name{nw_triangle}
\alias{nw_triangle}
\title{The NW Triangle Plot}
\usage{
nw_triangle(obj, add = FALSE, details = FALSE, ...)
}
\arguments{
\item{obj}{an object returned by nwbass function}

\item{add}{FALSE. Should data be added to an existing plot?}

\item{details}{FALSE. Should details about other distributions be included (must have ADD=FALSE)}
}
\value{
Creates the NW Triangle plot
}
\description{
A function for plotting the NW Triangle for an object created with nwbass()
}
\details{
Plots the posterior draws of the steepness and asymmetry parameter, defined as (1+gamma)^(-1/2) and beta/sqrt(gamma^2+beta^2)*steepness respectively.
These parameters are location and scale invariant. Normal (0,0) and Cauchy (0,1) occur as limiting cases.
}
\examples{
#not yet

}
