% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwbass.R
\name{nwbass}
\alias{nwbass}
\title{Generalized Bayesian MARS with a Normal Wald Likelihood}
\usage{
nwbass(
  X,
  y,
  w_prior = list(type = "GIG", p = 0, a = 0, b = 0, prop_sigma = 0.2),
  maxInt = 3,
  maxBasis = 1000,
  npart = NULL,
  nmcmc = 10000,
  nburn = 9001,
  thin = 1,
  moveProbs = rep(1/3, 3),
  a_tau = 1/2,
  b_tau = NULL,
  a_lambda = 1,
  b_lambda = 1,
  m_beta = 0,
  s_beta = 100,
  lag_beta = round(nmcmc * 0.5),
  m_gamma = 1,
  s_gamma = 0,
  vsig = 0.25,
  scale = 1,
  Iw0 = rep(1, maxInt),
  Zw0 = rep(1, ncol(X)),
  verbose = TRUE
)
}
\arguments{
\item{X}{an Nxp matrix of predictor variables.}

\item{y}{an Nx1 matrix (or vector) of response values.}

\item{w_prior}{a named list specifying the prior for the global variance component. See details.}

\item{maxInt}{integer for maximum degree of interaction in spline basis functions. Defaults to the number of predictors, which could result in overfitting.}

\item{maxBasis}{maximum number of basis functions. This should probably only be altered if you run out of memory.}

\item{npart}{of non-zero points in a basis function. If the response is functional, this refers only to the portion of the basis function coming from the non-functional predictors. Defaults to 20 or 0.1 times the number of observations, whichever is smaller.}

\item{nmcmc}{number of mcmc iterations}

\item{nburn}{number of burn-in samples}

\item{thin}{thinning for mcmc}

\item{moveProbs}{a vector defining the probabilities for (i) birth (ii) death and (iii) mutation. Default is rep(1/3,3).}

\item{a_tau}{prior for tau}

\item{b_tau}{prior for tyau}

\item{a_lambda}{prior for lambda}

\item{b_lambda}{prior for lambda}

\item{m_beta}{prior for beta}

\item{s_beta}{prior for beta}

\item{lag_beta}{number of steps for which beta is fixed to m_beta (usually zero)}

\item{m_gamma}{prior for gamma}

\item{s_gamma}{prior for gamma}

\item{vsig}{proposal sd for v_i}

\item{scale}{fixed variance parameter. default is one.}

\item{Iw0}{vector of nominal weights for degree of interaction, used in generating candidate basis functions. Should have length equal to Jmax and have positive entries.}

\item{Zw0}{vector of nominal weights for variable selection, used in generating candidate basis functions. Should have length equal to ncol(X) and have positive entries.}

\item{verbose}{Logical. Should gbass print completion status? Default TRUE}

\item{v_prior}{a named list specifying the (shared) prior for the local variance components. See details.}
}
\value{
The returned value is a named list with components for each of the MCMC parameters. The acceptance rates for each move type is returned. If applicable, we also return acceptance rates for w and the v_i.
}
\description{
Fits a generalized BMARS model with Normal-Wald likelihood. General purpose regression with flexible error distribution (unimodal).
}
\details{
See details for gbass(). Additionally, a normal prior must be specified for asymmetry parameter (beta) and tail heaviness parameter (gamma)
See also nw_gamma_prior() and nw_triangle()
}
\examples{
n <- 100 #Number of observations
p <- 4   #Number of variables (beyond p = 2, variables are inert)
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, ff1)
y <- y + rgamma(n, 1, 0.1)/4
gamma_prior <- nw_gamma_prior(p1=0.5, p2=1e-4) #Get reccomended prior for gamma
mod <- nwbass(X, y, m_beta=0, s_beta=30, m_gamma=gamma_prior[1], s_gamma=gamma_prior[2],
              nmcmc=1000, nburn=901)
nw_triangle(mod)

}
