% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbass.R
\name{plot.gbass}
\alias{plot.gbass}
\title{Plot method for class "gbass"}
\usage{
\method{plot}{gbass}(x, ...)
}
\arguments{
\item{x}{an object of class "gbass" created by \code{gbass}, \code{nwbass}, or a wrapper.}
}
\description{
This function plots an object of class "gbass"
}
\details{
Returns a matrix of posterior predictions.
}
