% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gbass}
\alias{plot.gbass}
\title{Plot function for class "gbass"}
\usage{
\method{plot}{gbass}(x, ...)
}
\arguments{
\item{object}{an object of class "gbass" created by the gbass, tbass or hbass functions}

\item{newdata}{a matrix of predictor variables with ncol(newdata) == ncol(X)}

\item{mcmc.use}{a vector subsetting which posterior samples to use for prediction. Default is to retain all samples.}
}
\description{
This function plots an object of class "gbass"
}
\details{
Returns a matrix of posterior predictions.
}
