% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gbass}
\alias{predict.gbass}
\title{Predict function for class "gbass"}
\usage{
\method{predict}{gbass}(object, newdata = NULL, mcmc.use = NULL)
}
\arguments{
\item{object}{an object of class "gbass" created by the gbass, tbass or hbass functions}

\item{newdata}{a matrix of predictor variables with ncol(newdata) == ncol(X)}

\item{mcmc.use}{a vector subsetting which posterior samples to use for prediction. Default is to retain all samples.}
}
\description{
This function generates posterior predictions using predictors Xnew and a model of class "gbmars"
}
\details{
Returns a matrix of posterior predictions.
}
