% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMHN.R
\name{rMHN}
\alias{rMHN}
\title{rMHN(n=10000,alpha=32,beta=20,gamma=-3)}
\usage{
rMHN(n = 1, alpha, beta, gamma)
}
\description{
rMHN(n=10000,alpha=32,beta=20,gamma=-3)
}
\examples{
ss=rMHN(n=10000,alpha=32,beta=20,gamma=-3)
ss=rMHN(n=1000, alpha=3, beta=1, gamma=1)
ss=rMHN(n=1000, alpha=3, beta=1, gamma=-1)
hist(ss$sample,prob=TRUE, col="grey")
lines(density(ss$sample))
}
