% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tbass}
\alias{tbass}
\title{TBASS - Bayesian MARS with a Student's t likelihood}
\usage{
tbass(X, y, df = 5, ...)
}
\arguments{
\item{X}{an Nxp matrix of predictor variables.}

\item{y}{an Nx1 matrix (or vector) of response values.}

\item{df}{degrees of freedom, default is 5.}
}
\description{
TBASS - Bayesian MARS with a Student's t likelihood
}
\details{
Inverse-gamma prior for local variance components. If df > 2, then
the scale is set to (df-2)/df so that the variance of y is equal to w.
}
